third_party/WebKit/LayoutTests/crypto/subtle/aes-kw/wrap-unwrap-aes-expected.txt

Test wrapping and unwrapping keys with AES-KW.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Importing key encryption key...
Importing a key to be wrapped...
Wrapping it...
PASS bytesToHexString(wrappedKey) is '1fa68b0a8112b447aef34bd8fb5a7b829d3e862371d2cfe5'
Unwrapping it...
PASS unwrappedKey.toString() is '[object CryptoKey]'
PASS unwrappedKey.type is 'secret'
PASS unwrappedKey.extractable is true
PASS unwrappedKey.algorithm.name is 'AES-CBC'
PASS unwrappedKey.algorithm.length is 128
PASS unwrappedKey.usages is ['encrypt', 'decrypt', 'wrapKey', 'unwrapKey']
Exporting it...
PASS bytesToHexString(unwrappedKeyData) is bytesToHexString(keyData)
PASS successfullyParsed is true

TEST COMPLETE
